pyproject.toml
[project]
name = "ecosistema-datos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "contextily>=1.7.0",
    "coverage>=7.12.0",
    "earthpy>=0.9.4",
    "faiss-cpu>=1.13.0",
    "folium>=0.20.0",
    "geemap>=0.36.6",
    "geopandas>=1.1.1",
    "groq>=0.36.0",
    "h5py>=3.15.1",
    "harmony-py>=0.4.14",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "langchain>=1.1.0",
    "matplotlib>=3.10.7",
    "nltk>=3.9.2",
    "numpy>=2.3.5",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "papermill>=2.6.0",
    "pyarrow>=21.0.0",
    "pygbif>=0.6.6",
    "pystac-client>=0.9.0",
    "pytest>=9.0.1",
    "python-docx>=1.2.0",
    "python-dotenv>=0.20.0",
    "pyyaml>=6.0.3",
    "rasterio>=1.4.3",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "shapely>=2.1.2",
    "streamlit>=1.51.0",
    "streamlit-folium>=0.25.3",
]
